fix listCount return type

+++ CPP-A.01_GLDraw2D/inc/CCircle.h
@@ -21,7 +21,7 @@ public:
     void set(CPoint c, float r);
     void set(float x, float y, float r);
     void list();
-    static int listCount();
+    static unsigned long listCount();
 };
 
 #endif //CPP_CCIRCLE_H
