+++ CPP-A.01_GLDraw2D/inc/CRect.h
@@ -23,7 +23,7 @@ public:
     void set(CPoint tl, CPoint br);
     void set(float x1, float y1, float x2, float y2);
     void list();
-    static int listCount();
+    static unsigned long listCount();
 };
 
 #endif //CPP_CRECTANGLE_H
