+++ CPP-A.01_GLDraw2D/src/CRect.cpp
@@ -57,7 +57,7 @@ void CRect::list()
     cout << "]" << endl;
 }
 
-int CRect::listCount()
+unsigned long CRect::listCount()
 {
     return ulCount;
 }
